Allow user-defined mappings without clobbering defaults.
Keep the default mappings in a script-local dictionary, and allow the
user to safely extend such dictionary with more completion mappings, but
without permitting the user to override the defaults.

To keep things simple, when automatic completion is enabled, <c-p> is
appended to each user-defined mapping to avoid selecting the first entry
in the pop-up menu. This should work in most cases (in fact, the only
cases when this would not work as expected is when the user-defined
mapping by default selects an entry different from the first).