Allow the user to override the default can_complete.
Now g:mucomplete#can_complete is built so that 'default' entries set by
the user are merged into the default 'default' entry, and other entries
are added to the dictionary. So, if for example the user defines
(untested):

    let g:mucomplete#can_complete = {
      \  'default' : { 'file' : { t -> t =~# '[\:/]' } },
      \  'cpp' : { 'omni' : { t =~# '::$' } }
      \ }

then for all 'file' completions the condition set by the user will be
used; omnicompletion will be usually tried when the default condition is
satisfied, except for cpp filetypes, for which omnicompletion will be
tried only when the text matches the pattern set by the user; for all
other methods, the default conditions will be tested.