Fix problem with remappings.
Before this commit, there was a problem with completion mappings that
remapped to something else (like SQL <c-c> mappings), due to the fact
that in something like

    <c-r>="\<c-c>a"<cr>

`<c-c>a` is not expanded. So, for example, the following completion
chain was buggy:

    let g:mucomplete#user_mappings = {
        \ 'sqls' : ["\<c-c>s", "\<c-c>s\<c-p>"],
        \ 'sqlk' : ["\<c-c>k", "\<c-c>k\<c-p>"]
        \ }
    let g:mucomplete#chains = { 'sql' : ['file', 'sqls', 'sqlk', 'keyn'] }

When typing something like `ocr<tab> ta<tab>`, the first <tab> would
correctly complete the keyword `create` using <c-c>s, but the second
would fail. The second <tab> would be remapped to

    <c-c>s<c-r>=pumvisible()?'':mucomplete#complete_chain(1)<cr>

Since <c-c>s would not return any result, mucomplete#complete_chain(1)
would be called and <c-c>k fed to the buffer. But this latter mapping,
being evaluated inside <c-r>=, would not be remapped, causing Vim to
exit Insert mode and go up one line (because of `k`).

To fix that, I have introduced a new <plug>, to be used internally,
which is used by mucomplete#complete_chain() in place of the previous
recursive call.