Fix regression from c173570ac8.
Fix the nasty bug described in the Readme and discussed at
https://github.com/lifepillar/vim-mucomplete/issues/14.

Calling `pumvisible()` inside `mucomplete#complete_chain()` does not
return reliable results, but it seems that calling `pumvisible()` from
inside a `<c-r>=` expression does.

So, to fix the problem we use a `<c-r>=` expression to set an internal
variable that is subsequently used by `mucomplete#complete_chain()` to
determine that the chain should be stopped.