Define mappings only once.
The logic for switching state when the pop-up menu is visible has been
moved into `s:act_on_pumvisible()`. This makes the double definition of
mappings redundant.

It also makes it easier for users to add completion methods that do not
enter the ctrl-x submode (e.g., the UltiSnips method).