Implement cycling through the completion chain.
This is a refactoring that adds the possibility to cycle through the
completion chain. If the first method returning results is not the one
you want, you may press <c-l> while the pop-up menu is visible to try
the next method. You may keep pressing <c-l> to cycle through all the
completion methods and, similarly, use <c-h> to cycle backwards.

The code needs some clean-up, but it works.