Honour noinsert and noselect settings.
Fixes https://github.com/lifepillar/vim-mucomplete/issues/25.

Strictly speaking, completeopt's noinsert is honoured only when
automatic completion is on: for manual completion, an entry is always
inserted (this is the most reasonable behaviour).

Note that `g:mucomplete#auto_select` is not needed any longer: selecting
a menu entry is now the default behaviour (Vim's default behaviour),
which may be overridden with:

    set completeopt+=noselect

I have removed also `g:mucomplete#user_mappings_deselect`, because I
don't think that there will ever be a need for it.