Improve file/path completion.
When completion is done, check whether what was just completed was a
file/path. If that is the case, and the inserted text ends with a
(back)slash, trigger file/path completion again.

For 'file' completion, such behaviour is triggered after accepting a
menu entry with <c-y>. For 'path' completion, the behaviour is triggered
by pressing / (or backslash, in Windows).