Implement alternative spelling suggestions.
Vim's built-in spelling suggestion offers corrections for a word
*before* the cursor, but not necessarily the one *immediately before*,
so it's not very suitable for automatic completion, because it makes the
cursor jump.

This commit implements a new completion method, called 'uspl', which
still leverages Vim's features through the use of `z=`, but acts only on
the word immediately preceding the cursor.