Fix 'line' completion method inserting a new line.
Line completion seems to work differently from other completion methods:
typing a character that does not belong to an entry does not exit
completion. Before this commit, with autocompletion on such behaviour
resulted in µcomplete inserting a new line while the user was typing,
because µcomplete would insert <c-x><c-l> while in ctrl-x submode.

To fix that, we use the same trick as with 'c-p': make sure that we are
out of ctrl-x submode before typing <c-x><c-l>.