Allow CTRL-C to stop completion without messing up.
A completion mapping may start a long computation (e.g., thesaurus
completion), which the user may want to interrupt by pressing CTRL-C.

Before this commit, pressing CTRL-C while Vim is building the pop-up
menu causes

    =pumvisible()?mucomplete#yup():''

to be inserted literally in the buffer. It seems that pressing CTRL-C
causes the next character to be eaten (<c-r>, in this case). I don't know
exactly why that happens, but a workaround is to use <c-r><c-r> instead
of just <c-r> (after all, it doesn't make a difference for the
evaluation). The first <c-r> is then eaten, but the second allows the
expression to be evaluated correctly.

Btw, typing the sequence <c-x><c-t><c-c> after a word (that is,
initiating thesaurus completion, then interrupting it) causes the pop-up
menu to appear and stay open, presumably partially populated. For some
reason, though, with the workaround above the pop-up menu is dismissed
(almost) immediately. This is not ideal, but that's the best I can do
for now.