Add flag telling whether completion is started manually.
The flag can be used to enable certain completion methods only when the
user starts completion manually, or only when completion is triggered
automatically.

The use case I have in mind is the following. Consider a slow completion
method, say thesaurus completion. It doesn't make much sense to use
thesaurus completion with automatic completion, because it would be
called too frequently. Yet the user might still want to trigger
thesaurus completion by pressing <tab>. One possibility would be to
implement two default completion chains, one for manual completion and
one for automatic completion. But that would make both the interface and
the implementation too complicated.

Instead, we may use g:mucomplete_with_key to test whether completion is
initiated by the user or automatically. For instance:

    let g:mucomplete#can_complete = { 'default':
      \ 'thes': { t -> g:mucomplete_with_key && strlen(&l:thesaurus) > 0 }
    let g:mucomplete#chains = { 'default': ['keyn', 'thes'] }

This would enable the 'thes' method only when the user presses
<tab> or <s-tab>; for automatic completion, only keyword completion
would be used.