Set s:N as part of the script's status setup.
This fixes the problem reported in this pull request:

    https://github.com/lifepillar/vim-mucomplete/pull/35

The problem is that `mucomplete#cycle()` may be invoked before
`mucomplete#complete()` is called for the first time. Since, before this
commit, `s:N` was initialized in `mucomplete#complete()`, invoking
`mucomplete#cycle()` resulted in an `Undefined variable` error.

The pull request highlights a more general situation: a user may mix
µcomplete completion (automatic or with `<tab>`) with standard
completion mappings. In particular, a user may open a pop-up menu using
standard completion mappings, and then use `<c-h>` and `<c-l>` to cycle
through the current completion chain. We must make sure that such
interleaving of mappings does not result in bugs.

The fix in this commit establishes the following invariant: at any time,
`s:N` is the length of `s:compl_methods`, that is, the length of the
current completion chain. As far as I see, this is necessary and
(hopefully) sufficient to guarantee that the above mentioned
interleaving does not cause any problem.

There may still be (rare) situations in which the current completion
chain and the internal completion chain are not in sync, though. For
instance, if the user sets `b:mucomplete_chain` to a new value, then
completes using `<c-x><c-n>` (bypassing µcomplete) and finally presses
`<c-l>` while the pop-up menu is visible, µcomplete will still cycle
through the completion chain that was defined before setting
`b:mucomplete_chain`. I think that such misbehavior might be avoided by
adding a test in `mucomplete#cycle()`, checking for
`g:mucomplete_with_key || s:auto` (to be reasonably sure that the pop-up
menu has appeared as a result of µcomplete's action). I am not sure that
such a corner case is worth such a fix, though.