Check for uniqueness of mappings.
Suppose that the user defines:

    imap <c-j> <plug>(MUcompleteFwd)

in his vimrc. When he types <c-j>, that will be mapped to
<plug>(MUcompleteCycFwd) rather than <plug>(MUcompleteFwd), because that
is the default mapping when not overridden by the user. This may appear
as an unexpected behaviour.

By adding <unique>, we make sure that the mapping above raises an error,
unless <plug>(MUcompleteCycFwd) is also mapped to something else.