Do not trigger tab-completion with noexpandtab.
When you are editing a buffer with noet, you'd likely want <tab> to
insert a tabulation character even when the cursor is not immediately
preceded by a space. Before this commit, that was not possible
with µcomplete installed, unless the completion trigger was changed to something
different from <tab>.

This commit enables the following behaviour: if µcomplete's manual
completion trigger is <tab> (i.e., the default) and the current buffer
has noexpandtab, then pressing <tab> will always insert a tab character.
Autocompletion will behave as usual. If you want to manually complete
a word, you may either turn expandtab on or change µcomplete's
completion trigger to something different from <tab>.

When expandtab is on, everything behaves as it has always done.