Allow inverting cycling direction in pop-up menu.
This commit introduces a new setting:

    let g:mucomplete#popup_direction = { 'c-p' : -1, 'keyp' : -1 }

This is a hash mapping completion methods to values in {1, -1}.

The idea is to allow the user to choose, for each completion type,
whether the first selected entry should be the topmost one (1) or the
bottomost one (-1), and, consequently, whether pressing <tab> while the
pop-up menu is open should move top-to-bottom or bottom-to-top.

The default behaviour follows Vim's convention to always select the
topmost entry, except for ctrl-p completions ('c-p' and 'keyp' methods),
which select the last entry. Hence, the movement of <tab>/<s-tab> in the
pop-up menu is inverted for those completion methods (i.e., <tab> moves
from bottom to top and <s-tab> moves from top to bottom).

The default may be overridden by setting g:mucomplete#popup_direction in
your `vimrc`. For instance, if you want 'c-p' to be top-to-bottom and
'file' to be bottom-to-top, put this in your `vimrc`:

    let g:mucomplete#popup_direction = { 'c-p': 1, 'file': -1 }