Improve automatic completion.
This commit introduces a major change aiming at improving the user
experience when automatic completion is on. Now the pop-up menu should
not go on and off on alternating keys (as reported, for example, in
https://github.com/lifepillar/vim-mucomplete/issues/50).

I have simplified the code of s:act_on_textchanged(), which now triggers
completion again for all completion types, and not only for file/path
completion. I have adapted the trigger_auto_pattern accordingly.

To make all this work, unfortunately, the user is required to add a few
mappings in his vimrc. Defining thin wrappers around <c-e>, <c-y>, and
<cr> is the only way (as far as I know) to detect whether the pop-up
menu is dismissed using one of those keys. When the pop-up menu is
abandoned with <c-e> (or <c-y> or <cr>), completion is not triggered
again (that would be unexpected and, in fact, it was not working
correctly for file/path completion before this commit). If the pop-up
menu is closed because the user presses some other key (presumably
because he's going on typing) then completion is triggered again (if the
trigger_auto_pattern matches, of course).