Get rid of s:pumvisible.
Apparently, what is important when trying a completion is to use <c-r>
before invoking the plug. What I had not realized before is that even an
empty expression will do. So, there is no need for s:pumvisible.

Note that <c-r><c-r>=<cr> is necessary. Without that part, all
completion messes up.