Try to minimize side effects.
Here I try to combine the best of both worlds using InsertCharPre and
TextChangedI together. InsertCharPre has the following features:

- fired only when a (printable) character is inserted (so, for example,
  it is not invoked on backspace or CTRL-E);
- not triggered when 'paste' is on;
- triggered even when pumvisible() is true.

TextChangedI has the following features:

- fired with every change, even when the pop-up menu is dismissed with
  CTRL-E, CTRL-Y, or Enter (even with no change at all? See
  https://github.com/vim/vim/issues/679);
- not triggered when the pop-up menu is visible.

MUcomplete does not need to do anything special while the pop-up menu is
visible. But I'd prefer not to have auto-completion triggered on
backspace, and of course when 'paste' is on MUcomplete shouldn't do
anything. Besides, when completion takes relatively long, if it is
invoked before inserting a character (on InsertCharPre), the user may
notice a slight delay in character insertion. Not good.

So, here how I do combine the two events to get the best of both worlds:
on InsertCharPre, MUcomplete just records that a key was inserted. All
the work is performed on TextChangedI, but only if the TextChangedI
event is preceded by InsertCharPre (which MUcomplete knows through
s:insertchapre). By doing so, auto-completion is not triggered when the
user uses backspace and MUcomplete does nothing as long as the pop-up
menu is visible (because TextChangedI is not fired then). When the
pop-up menu is dismissed because the user inserts a character,
InsertCharPre is fired before the character is inserted (and when the
pop-up menu is still visible), then TextChangedI is fired (after the
character has been inserted and the pop-up menu has disappeared). Since,
s:insertcharpre is 1 at that point, auto-completion kicks in again, as
desired.

Since s:act_on_textchanged() is now invoked only when pumvisible() is
false, we do not need to prepend the plug with s:ctrlx_out any more,
which seems to help with (although possibly doesn't solve completely) an
annoying artefact in the terminal (at least in macOS's Terminal.app): if
the cursor is set to a bar in Insert mode and to a block in Normal mode,
the cursor may flicker (i.e., it may become block-wise intermittently) while
typing in Insert mode, as if Vim switched briefly to Normal mode (which
might happen indeed, since we call feedkeys()...).

Ok, this seems all good, but TextChangedI had been replaced in the first
place because of the cursor disappearing problem mentioned in
7d5fe661178989f9ec9aaca5f6d8cebcbff99cf9. To mitigate (and hopefully
solve) that problem, we now call feedkeys() only when at least one
completion method can be applied for sure. It seems to be an effective
strategy.

The code still needs some clean-up, but hopefully I've finally nailed
it.