Add option to trigger tab-completion unconditionally.
Normally, when the character before the cursor is a space or the cursor
is at the start of a line, pressing <tab> causes µcomplete to insert
a literal Tab character. There are situations, however, when it is
desirable to fire the completion chain even in such cases (for instance,
in Ledger buffers omni-completion triggered at the start of the line
inserts the current date).

This is now supported by the g:mucomplete#empty_text flag (and its
buffer-specific variant), which is assumed to be false by default.