Make tab-completion consistent when auto is on.
Even when auto-completion is turned on, it is still possible to trigger
the completion manually by pressing <tab>/<s-tab>. Before this commit,
however, pressing <tab> when auto was on did not ignore
noselect/noinsert. This commit fixes this inconsistency: now, when you
press <tab> or <s-tab> the first match is *always* automatically
inserted, no matter whether auto-completion is on.