Update troubleshooting vimrc.
Enable filetype detection and syntax highlighting.
Instead of an empty `runtimepath`, set it to `$VIMRUNTIME` (so, for
example, Vim's omni-completion functions can be used).

I have noticed that omni-completion may not work if you run

    vim -u troubleshooting_vimrc

inside µcomplete's directory (no idea why), but it works if
troubleshooting_vimrc is copied into ~/.vim. Make that explicit in the
documentation.