Reduce startup time.
At the expense of a bit of of code duplication, this commit speeds up
µcomplete's startup significantly when the user has set
g:mucomplete#enable_auto_at_startup in his or her vimrc.

Defining the augroup in plugin/mucomplete.vim is much faster than
calling :MUcompleteAutoOn, likely because the latter causes Vim to
search several locations for the file containing the function used by
MUcompleteAutoOn.

To give an idea, in my laptop, µcomplete would load in over 0.600ms
(with autocompletion on), while now it loads in less than 0.180ms.