Define g:mucomplete#cycle_all.
Addresses part of issue
https://github.com/lifepillar/vim-mucomplete/issues/76.

The problem is that, if there are methods that are used only when
completion is triggered manually (as in the thesaurus example in the
tips section of the documentation), cycling through the completion chain
ignores such methods when the pop-up menu appears automatically (i.e.,
when auto-completion is on). By setting this option to 1, all methods
are cycled through in any case.

So, in the thesaurus example, during normal typing only keyword
completion is active. To try searching for synonyms, you need to press
Tab or Shift-Tab. When the pop-up menu appears during typing and you
press CTRL-H or CTRL-J no cycling occurs, because thesaurus completion
is skipped by default. But if you set g:mucomplete#cycle_all to 1, then
thesaurus completion is considered when cycling, even if the completion
menu has been triggered automatically and not by a key.