Initialize s:compl_methods with a non-empty chain.
This is to fix a minor bug that could be triggered as follows before
this commit:

1. Open Vim
2. Type something like `foobar fo<c-p>`
3. Press <tab>

Since <tab> is pressed for the first time when the pop-up menu is
already visible, the first call to mucomplete#tab_complete() results in
mucomplete#cycle_or_select() to be invoked with s:compl_methods being
empty. By initializing s:compl_methods with a non-empty completion
chain, access to s:compl_methods in that circumstance becomes valid.