Implement look-behind for the text to be completed.
This change fixes an issue with completing paths with spaces.
Before this commit, if you had a path like

/abc/foobar xyz

(where `foobar xyz` is the name of a directory) and the buffer contained

/abc/foobar x

with the cursor just after x, pressing <tab> would not trigger path
completion. That happened because µcomplete looked at one word before 
the cursor (in the case, just `x`) and would not recognize that word as
part of a potential path.

In general, to determine that the text before the cursor may be a path,
µcomplete must find a path separator or a tilde. In the worst case,
i.e., when the current line does not contain any path separators or
tilde characters, this takes time proportional to the length of the
line.

To keep things fast, µcomplete now examines only a bounded number of characters (30 by default). This does not change µcomplete's performance
(in fact, with this commit, completion may even be a few µs faster), but
it makes path completion work better.

A new setting, mucomplete#look_behind, can be used to adjust the
look-behind length.