Fix issue with cycling when auto-completion is on.
This commit fixes an issue with cycling not offering alternative
completion methods in some cases when auto-completion is on.
For example, suppose that the buffer contains:

  autobahn

and the user starts typing

  ./

This text will trigger path completion. Assume that the pop-up menu
appears offering some directories to complete, among which `autoload`.
Note that, at this point, the value of s:compl_text is './'.
Now the user keeps typing:

  ./aut

Since the pop-up menu is visible, µcomplete does nothing (Vim will
restrict  the menu choices to those matching 'aut', of course). Now,
the user types CTRL-J. In principle, this should switch to the next
available completion method, which by default is keyword completion.
So, the user would expect the pop-up menu to change and offer
'autobahn'. But, since s:compl_text is still './', µcomplete will not
find and keyword matching the text.

With this commit, s:compl_text is updated each time the user starts
cycling.