Fix regression preventing tabbing after spaces.
If the character in front of the cursor is a space (or if there is no
character in front of the cursor), the obvious action when the user
presses <tab> is to send a tab character, not to trigger completion.

The recent commits have broken this behaviour. This commit fixes it.