Escape special chars in paths when globbing.
This makes path expansion more robust, especially in Windows, where in
some cases errors would occur because some characters that are treated
specially by glob() are in isfname. In particular, this commit should
fix https://github.com/lifepillar/vim-mucomplete/issues/87.

There is still an issue when a path contains asterisks, which I think is
a bug in Vim. Say l:prefix is '~/foo/a*' (where 'a*' is the name of
a directory). Then glob() will receive fnameescape(l:prefix).'*', that
is, '~/foo/a\**'. As of Vim 8.0.1496, Vim expands the path as if glob()
had received '~/foo/a**' (thus returning the directory and its
subdirectories), i.e., the asterisk's escaping is ignored. See also
https://github.com/vim/vim/issues/2632.