Fix cycling mistakenly turning s:cancel_auto to 1.
Since <c-e> is mapped, using it in mucomplete#cycle() is wrong, because
it causes mucomplete#popup_exit() to be invoked and, consequently,
s:cancel_auto to be set to 1. That means that, if the user keeps typing
after cycling and the popup menu is dismissed because there is no match
for the current method, other methods are not attempted immediately.
For example, in a buffer with 'spell' on and automatic completion on,
type:

foobar
foo<c-j>

At this point, the popup menu shows spelling corrections. If the user
types 'b', the menu is dismissed because there is no match for 'foob'.
Before this commit, the user would have to type 'a' before µcomplete
proposes 'foobar' as a keyword completion. After this commit, keyword
completion is attempted immediately (when the user types 'b').