Add (failing) test for path completion and double-slash comments.
MUcomplete triggers path completion in situations when it shouldn't.
For example, if the current directory contains a file called 'foo`,
and the user types something like:

// f

then presses <tab>, MUcomplete offers 'foo' as a path completion.
That happens because the line contains slashes, so it potentially
describes a path, and MUcomplete tries to glob, in order, the
string '// f' (with no result), and 'f' (finding 'foo').