Get rid of pop-up mappings. Remove smart enter option.
I have finally found a way to re-trigger completion when the pop-up menu
is dismissed with a key different from <c-e>, <c-y> or <cr>, which does
not involve remapping those keys. I have realised that, when that
happens, Vim stays in Insert mode completion, which can be tested with
mode(1).

Getting rid of <c-e>, <c-y> and <cr> mappings avoids an entire class of
conflicts with other plugins (e.g., auto-pairs, endwise).

The only drawback (AFAICS) is that the smart enter option (which was
based on remapping <cr>) does not work any longer, hence it is removed,
too. This is not a big deal, as the user can add a mapping in her or his
vimrc (this is explained in :help mucomplete-tips). This gives the user
much more freedom in defining mappings (g:mucomplete#smart_enter was not
a good idea to start with, but it made sense before as I had to remap
<cr> anyway).

This commit should also prepare MUcomplete for an interesting
enhancement: delayed autocompletion, i.e., autocompletion triggered
after 'updatetime' ms (see https://github.com/lifepillar/vim-mucomplete/issues/107).

All tests still pass (I have removed those related to the code that has
been deleted, of course).