Restore backward compatibility with Vim <8.0.0283.
Maintaining a different branch for older Vim is not a good idea either
from a developer's perspective or for users. This refactoring brings
back compatibility with older Vim to the master branch.

This commit also enables delayed autocompletion in Vim older than
v8.0.0283. It works nearly exactly as in newer Vim, with the exception
that backspace triggers completion immediately. This is due to the fact
that MUcomplete cannot accurately detect when Vim in in Insert
completion mode without patch 8.0.0283. Maybe I'll find some clever
trick in the future...