Improve backward compatibility for autocompletion.
Now autocompletion works with Vim 7.4.143 or later.

In versions less than 7.4.775, there is no equivalent of noinsert (i.e.,
select the first menu entry without inserting it): MUcomplete always
behaves as if noselect were added to completeopt. It would be easy to
add an option to implement noinsert in older Vim versions, though.