Set s:compl_text after <c-e> when cycling.
Fix a subtle bug I had stumbled upon while editing C files (see test).
When the buffer contained `int` and I was typing `in<tab>`, MUcomplete
would offer `incCount(` as an omni-completion (such function was defined
in a tags file). Now, cycling with <c-h> or <c-j> would dismiss the
menu; instead, I would have expected to switch to keyword completion and
offer `int` as a completion.

The reason is that s:compl_text was set too early in mucomplete#cycle(),
i.e., before the menu was dismissed by <plug>(MUcompleteCte).

In fact, the text that s:next_method() (hence, s:can_complete()) would
use was the current text in the buffer (i.e., `incCount(` in the example
above) and not the text typed by the user (`in` in the example above).

This commit fixes the inconsistency and sets the correct order of
operations:

1. the menu choice is cancelled with <c-e> (via <plug>(MUcompleteCte));
2. s:compl_text is set to the text in the buffer (via new
   <plug>(MUcompleteCyc)).

This also deals correctly with situations such as the following (see
test):

1. the user types something and triggers completion, say `in<tab>`;
2. a completion method is successful and the menu appears;
3. the user unselects the current choice (e.g., by pressing <c-p>) and
   keeps typing characters, say `d`, which do *not* cause the menu to be
   dismissed (because the text still matches some menu entry);
4. the user selects some menu entry;
5. the user activates cycling via <c-h> or <c-j>.

At this point, the text that should be used is *not* `in` and *not* the
currently selected menu entry, but `ind`. Before this commit, the
currently selected menu entry would be used.

Another way to see it is, <c-h> and <c-j> mean: "cancel the current
menu (as if typing <c-e>) and try completing my original text in a
different way".