Try avoiding false positives when matching paths.
Make path completion rules stricter when automatic completion is on.
In particular, with this commit path completion is not triggered (or
returns nothing) when:

1. the user types a path separator (e.g., /);
2. the user types two consecutive path separators (e.g., //).

Rule 2 means that typing comments in C/C++ now does not trigger path
completion any longer.

Nothing changes for manual completion.

The 'file' method is left unchanged.