Add an option to delay completion with timers.
This addresses
https://github.com/lifepillar/vim-mucomplete/issues/107 and is
essentially based on code proposed by @kristijanhusak for
https://github.com/lifepillar/vim-mucomplete/pull/114.

The feature mostly works, but has an annoying issue when using Vim
with patch 8.0.0283. Suppose that auto-completion is on and that
g:mucomplete#completion_delay is set to 1000. In a buffer containing
`foobar`, quickly type `foo<tab>` (i.e., trigger completion manually
before it kicks in automatically). Then, type CTRL-E to cancel your
menu choice. Then wait. After 1s, the menu will pop up again
automatically.

This happens because after the menu is dismissed with CTRL-E,
TextChangedI is triggered, which calls mucomplete#auto#complete().
With CursorHoldI, we used mucomplete#auto#ic_auto_complete() to
decide whether auto-completion should be triggered again, but it is
not immediately obvious how such function could be integrated into
the new method based on timers.