Add support for notifications (idea from @mg979).
Define a user autocommand that is triggered upon successful completion.

Use the autocommand to notify which completion method is used by
MUcomplete in the active pop-up menu.

Define a command to set the notification level (~verbosity).

Notifications can be displayed in the command line (level 1 or 2) or in
the status line (level 3).

The code in this commit is adapted from @mg979's pull request:

https://github.com/lifepillar/vim-mucomplete/pull/118