With delay on, apply delay to auto-reopen.
When the pop-up menu is open, the user may keep typing until the typed
text does not match any menu item any more. In such cases, Vim dismisses
the pop-up menu, of course. If auto-completion is on, MUcomplete retries
completion immediately, because other completion methods may be applied.

For example, suppose that the buffer contains the word 'sure' and that
the current completion chain has keyword completion and spelling
completion. Assume that 'spell' is on and that the user starts typing
new text. After typing:

su

MUcomplete triggers keyword completion and offers 'sure' in the pop-up
menu. Now, suppose that the user keeps typing:

surel

When 'l' is typed, the pop-up menu is dismissed. But MUcomplete is
clever enough to trigger completion again, and in fact, in this case it
offers spelling suggestions.

Before this commit, auto-completion is always retried *at once* after
the pop-up menu is closed as explained above, even if the user has set
g:mucomplete#completion_delay to a positive value.

With this commit, MUcomplete checks whether
g:mucomplete#completion_delay is set to a positive value, in which case
it waits the amount of time specified by such setting before retrying
auto-completion.

The main reason for introducing this behaviour is to address issue
https://github.com/lifepillar/vim-mucomplete/issues/139. There are
apparently some timing/synchronization issues with the LanguageClient
plugin, which delayed completion seems to mitigate. But that requires
a small delay to be applied also in the circumstances described above.

WARNING: this new feature has been tested very lightly, and *only* with
timers.

NOTE: in a future commit, I plan to introduce a separate option to
trigger this behaviour.