Do not break delayed completion with updatetime.
This change restores the correct behaviour of delayed completion when
g:mucomplete#completion_delay is set to 1 (hence, the delay is based on
updatetime and completion is triggered by a CursorHoldI event).

Note that the feature added in the previous commit does not apply to
completion delayed with CursorHoldI, that is, when
g:mucomplete#completion_delay = 1, completion is always triggered immediately after popup close and auto-reopen.

The feature introduced in the previous commit cannot be extended to this
kind of delayed completion because CursorHoldI is not triggered in
Insert completion mode.