Reuse mucomplete#chains for scoped chains.
Do not introduce a new variable for scoped chains; let
g:mucomplete#chains accept Dictionary as well as List values. For
instance, with this commit, instead of writing:

    let g:mucomplete#scoped_chains = {
          \ 'vim': {'vimString':      [],
          \         'vimLineComment': ['keyn'],
          \         'vimComment':     ['spel']
          \ }

one may write:

    let g:mucomplete#chains = {
          \ 'vim': {'vimString':      [],
          \         'vimLineComment': ['keyn'],
          \         'vimComment':     ['spel'],
          \         'default': ['path', 'cmd', 'keyn']
          \ }

I.e., g:mucomplete#chains accepts also Dictionaries as values. These map
syntax highlight items or the 'default' key to Lists of completion
methods. When 'default' is not present in the inner Dictionary, then MUcomplete falls back to the default chain
(`g:mucomplete#chains['default']`), which is always defined.

An advantage of this approach over the previous one is that it preserves
the symmetry with `b:mucomplete_chain`. One might define:

    let b:mucomplete_chain = {'vimString':      [],
          \         'vimLineComment': ['keyn'],
          \         'vimComment':     ['spel'],
          \         'default': ['path', 'cmd', 'keyn']
          \ }

in a Vim buffer, for example. Btw, in this specific example, in a Vim
buffer the following would be equivalent to the previous example:

    let b:mucomplete_chain = {'vimString':      [],
          \         'vimLineComment': ['keyn'],
          \         'vimComment':     ['spel']
          \ }

because MUcomplete already defines a filetype-specific completion chain
for Vim buffers.

NOTE: this has not been thoroughly tested yet.